+++ requirements_torch.txt
@@ -1,6 +1,8 @@
 --find-links https://download.pytorch.org/whl/torch_stable.html
-torch==1.8.1+cpu
-torchvision==0.9.1+cpu 
+#torch==1.8.1+cpu
+#torchvision==0.9.1+cpu 
+torch==1.10.1+cpu
+torchvision==0.11.2+cpu
 #pytorch_lightning==1.4.4
 
 onnxruntime==1.7.0
