+++ settings.json
@@ -1,1 +1,1 @@
-{"active_model": "20200430_003x_WMS", "exmask_enabled": false, "exmask_model": "20210629_014a_tape"}
+{"active_model": "022_segmentation.full", "exmask_enabled": false, "exmask_model": "20210629_014a_tape", "tracking_model": "022_roottracking.stage2"}
